New pre-built dataset impl -no forcefull and automatic update both for configuration and built in databases. User should trigger explicitly -Message will be shown to user when there is new biobtree-conf release -when --keep is not passed both index and db directory will be cleared before processing except during generate json files will remain in the index directory by deleting only *.gz files. -install command will clear everything and if built in database specified it will install and extract it as well.

+++ scripts/release.sh
@@ -10,9 +10,13 @@ elif [[ "$OSTYPE" == "darwin"* ]]; then
     OS="MacOS"
 fi
 
-rm biobtree 
+if [ -f biobtree ]; then
+    rm biobtree
+fi
 
-rm -rf biobtree_*.tar.gz 
+if [ -f biobtree_*.tar.gz ]; then
+    rm -rf biobtree_*.tar.gz
+fi
 
 go build
